[Decomp][Spell] Cohesion split: extract SpellPackets handlers to SpellPackets.cpp
